[ETW Exporter] Add Virtual destructor for TailSampler, Update Maintainer mode warnings for MSVC (#1897)

+++ examples/http/client.cc
@@ -82,7 +82,7 @@ int main(int argc, char *argv[])
   // The port the validation service listens to can be specified via the command line.
   if (argc > 1)
   {
-    port = atoi(argv[1]);
+    port = (uint16_t)(atoi(argv[1]));
   }
   else
   {
